refactor: billing & payment steps
